fix: fix the linked iterator will be invalid
When the std::unordered_multimap is rehashed, the iterators to the
elements in the map will be invalidated. If we want to traverse the
elements by the inserted order, we need to ensure that the iterators are
not invalidated.
This change use the std::pair<std::string_view, std::string_view> to
inserted of std::unordered_multimap::iterator. This will ensure that
the traverse behavior is not affected by the rehashing of the map.