feat(rule): improve matched variable handling
Add is_need_push_matched_ flag for Rule. If any action that requires the
matched variable is defined or MATCHED_VAR/MATCHED_VAR_NAME variable is
used, this flag will be set to true. Otherwise, it will be false.
This flag is used to optimize the performance of the rule evaluation.