feat: support unmatched/always actions in rules (#97)
The original implementation only evaluated actions when a rule matched.
This commit introduces support for evaluating actions even when a rule
does not match. Actions can now be categorized into 'matched' and
'unmatched' branches.
And the chained rules are evaluated based on whether the parent rule
matched or not, controlled by `!chain` or `*chain` actions.